add time translation

+++ app/src/kill.c
@@ -3,6 +3,16 @@
 
 int main(int argc, char *argv[])
 {
+	if(argc == 1) {
+		printf("kill: missing signal\n");
+		return -1;
+	}
+
+	if(argc == 2) {
+		printf("kill: missing process id\n");
+		return -1;
+	}
+
 	int sig = atoi(argv[1]);
 	int pid = atoi(argv[2]);
 	int ret = kill(pid, sig);
